# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "StoreFy API",
        "description": "StoreFY Swagger OpenApi description",
        "contact": {
            "email": "georgeton.ufg@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Api local server"
        },
        {
            "url": "http://storefy.workingtech.com.br:31080/api",
            "description": "Api dev server"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Login",
                "operationId": "Login",
                "requestBody": {
                    "description": "Logar no sistema",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "title": "email",
                                        "description": "email"
                                    },
                                    "password": {
                                        "title": "password",
                                        "description": "password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "title": "token",
                                            "description": "token",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "$ref": "#/components/schemas/validate_message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/bancos": {
            "get": {
                "tags": [
                    "bancos"
                ],
                "summary": "Get all items",
                "operationId": "getBancos",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/fields"
                    },
                    {
                        "$ref": "#/components/parameters/where"
                    },
                    {
                        "$ref": "#/components/parameters/orderBy"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "properties": {
                                                "id": {
                                                    "title": "id",
                                                    "description": "id",
                                                    "type": "integer"
                                                }
                                            }
                                        },
                                        {
                                            "$ref": "#/components/schemas/Bancos"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "bancos"
                ],
                "summary": "Add a new item",
                "operationId": "addBancos",
                "requestBody": {
                    "description": "Object to add new item",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Bancos"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "properties": {
                                                "id": {
                                                    "title": "id",
                                                    "description": "id",
                                                    "type": "integer"
                                                }
                                            }
                                        },
                                        {
                                            "$ref": "#/components/schemas/Bancos"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "$ref": "#/components/schemas/validate_message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/bancos/{uuid}": {
            "get": {
                "tags": [
                    "bancos"
                ],
                "summary": "Get item by ID",
                "operationId": "getBancosById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "$ref": "#/components/parameters/fields"
                    },
                    {
                        "$ref": "#/components/parameters/where"
                    },
                    {
                        "$ref": "#/components/parameters/orderBy"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "properties": {
                                                "id": {
                                                    "title": "id",
                                                    "description": "id",
                                                    "type": "integer"
                                                }
                                            }
                                        },
                                        {
                                            "$ref": "#/components/schemas/Bancos"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "bancos"
                ],
                "summary": "Update item",
                "operationId": "updateBancos",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Object to update item",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Bancos"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "properties": {
                                                "id": {
                                                    "title": "id",
                                                    "description": "id",
                                                    "type": "integer"
                                                }
                                            }
                                        },
                                        {
                                            "$ref": "#/components/schemas/Bancos"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "bancos"
                ],
                "summary": "Delete item",
                "operationId": "deleteBancos",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Item not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "validate_message": {
                "title": "Error message",
                "description": "Error jso1n",
                "required": [
                    "code",
                    "message"
                ],
                "properties": {
                    "code": {
                        "type": "string",
                        "readOnly": true
                    },
                    "message": {
                        "type": "string",
                        "minLength": 8,
                        "readOnly": true
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ErrorMessage"
                }
            },
            "Bancos": {
                "title": "Bancos model",
                "description": "Bancos model",
                "required": [],
                "properties": {
                    "field name": {
                        "title": "field title",
                        "description": "Field name"
                    },
                    "status": {
                        "title": "status",
                        "description": "Status",
                        "type": "boolean",
                        "default": false
                    },
                    "uuid": {
                        "title": "uuid",
                        "description": "Uuid"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Bancos"
                }
            },
...
        },
        "parameters": {
            "fields": {
                "name": "fields",
                "in": "query",
                "description": "Fields to return"
            },
            "where": {
                "name": "where",
                "in": "query",
                "description": "filter of any database fields (always send ternaries: 'field,operator,value')'"
            },
            "orderBy": {
                "name": "orderBy",
                "in": "query",
                "description": "Order ny database fields (always send: 'field,order')'"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}